added debugging levels
Instead of debugging being on or off I included debugging levels and a button that cycles through them.
      -- 0: debugging off
      -- 1: specific function debugging
      -- 2: infrequent function calls
      -- 3: include frequent function calls
      -- 4: display everything

It allows contributors to control the rate of debug messages sent to the Atom console so they can see what is going on.  Level 1 is practically the same as filtering by the function name, so no debug messages are set at level 1 right now.

Also added a printTable function useful for debugging.  It will print the contents of tables and any other types of values.