changing the structure of the puzzle
Keep two variables for a group of joined pieces.
"parentId":The pieceId of the parent piece that is still on the board.
"joinedTo":a circularly linked list of all the pieces that make up the group.

eg. if two pieces 45 and 87 are connected then
gameState.pieces[45]={joinedTo:87,parentId=45,player_name="jack"}
and
gameState.pieces[87]={joinedTo:45,parentId=45,player_name="jack"}

Since the "joinedTo" value points to the id of the next piece in the chain, these values can easily be corrupted, but I took care not to mess up the code for this.

New conventions, if the piece is unjoined, then parentId == joinedTo == nil. Also the parent id of the parent piece is its own id, instead of nil.

This implementation includes an iterator "iterJoinedTo". It takes the piece id of any piece and iterates through its own id and all the ids of the attached pieces.
for id in iterJoinedTo(45) do
print(id)
end
-- output is "45" then "87" then stops
To exclude itself, you can use iterJoinedTo(45,false) and it will only output "87" and stop.

The JSON save is updated.

It also assigns a name to both target piece and dropped piece if the name is blank, so that it will ping all pieces without any gaps.